+++ conan/api/subapi/profiles.py
@@ -3,7 +3,7 @@ import os
 from conan.api.output import ConanOutput
 from conan.internal.cache.home_paths import HomePaths
 
-from conans.client.loader import load_python_file
+from conan.internal.loader import load_python_file
 from conan.internal.api.profile.profile_loader import ProfileLoader
 from conan.internal.errors import scoped_traceback
 from conan.errors import ConanException
