+++ conan/api/subapi/upload.py
@@ -6,7 +6,7 @@ from conan.api.output import ConanOutput
 from conan.internal.conan_app import ConanApp
 from conan.internal.api.uploader import PackagePreparator, UploadExecutor, UploadUpstreamChecker, \
     gather_metadata
-from conans.client.pkg_sign import PkgSignaturesPlugin
+from conan.internal.rest.pkg_sign import PkgSignaturesPlugin
 from conan.internal.rest.file_uploader import FileUploader
 from conan.internal.errors import AuthenticationException, ForbiddenException
 from conan.errors import ConanException
