+++ conan/internal/api/profile/profile_loader.py
@@ -410,7 +410,7 @@ def _profile_parse_args(settings, options, conf):
 
 
 def migrate_profile_plugin(cache_folder):
-    from conans.client.migrations import update_file
+    from conan.internal.api.migrations import update_file
 
     profile_plugin_file = HomePaths(cache_folder).profile_plugin_path
     update_file(profile_plugin_file, _default_profile_plugin)
