+++ conan/internal/default_settings.py
@@ -169,7 +169,7 @@ build_type: [null, Debug, Release, RelWithDebInfo, MinSizeRel]
 
 
 def migrate_settings_file(cache_folder):
-    from conans.client.migrations import update_file
+    from conan.internal.api.migrations import update_file
 
     settings_path = os.path.join(cache_folder, "settings.yml")
     update_file(settings_path, default_settings_yml)
