+++ conan/internal/deploy.py
@@ -5,7 +5,7 @@ import fnmatch
 
 from conan.internal.cache.home_paths import HomePaths
 from conan.api.output import ConanOutput
-from conans.client.loader import load_python_file
+from conan.internal.loader import load_python_file
 from conan.internal.errors import conanfile_exception_formatter
 from conan.errors import ConanException
 from conans.util.files import rmdir, mkdir
