+++ conan/internal/graph/installer.py
@@ -8,7 +8,7 @@ from conan.internal.api.install.generators import write_generators
 from conan.internal.graph.graph import BINARY_BUILD, BINARY_CACHE, BINARY_DOWNLOAD, BINARY_EDITABLE, \
     BINARY_UPDATE, BINARY_EDITABLE_BUILD, BINARY_SKIP
 from conan.internal.graph.install_graph import InstallGraph
-from conans.client.source import retrieve_exports_sources, config_source
+from conan.internal.source import retrieve_exports_sources, config_source
 from conan.internal.errors import conanfile_remove_attr, conanfile_exception_formatter
 from conan.errors import ConanException
 from conan.internal.model.cpp_info import CppInfo, MockInfoProperty
