+++ conan/internal/loader.py
@@ -15,7 +15,6 @@ from pathlib import Path
 from conan.tools.cmake import cmake_layout
 from conan.tools.google import bazel_layout
 from conan.tools.microsoft import vs_layout
-from conans.client.loader_txt import ConanFileTextLoader
 from conan.internal.errors import conanfile_exception_formatter, NotFoundException
 from conan.errors import ConanException
 from conan.internal.model.conan_file import ConanFile
@@ -23,6 +22,7 @@ from conan.internal.model.options import Options
 from conan.api.model import RecipeReference
 from conan.internal.paths import DATA_YML
 from conan.internal.model.version_range import validate_conan_version
+from conans.util.config_parser import ConfigParser
 from conans.util.files import load, chdir, load_user_encoded
 
 
@@ -397,3 +397,54 @@ def _get_required_conan_version_without_loading(conan_file_path):
         pass
 
     return txt_version
+
+
+class ConanFileTextLoader:
+    """Parse a conanfile.txt file"""
+
+    def __init__(self, input_text):
+        # Prefer composition over inheritance, the __getattr__ was breaking things
+        self._config_parser = ConfigParser(input_text,  ["requires", "generators", "options",
+                                                         "imports", "tool_requires", "test_requires",
+                                                         "layout"],
+                                           strip_comments=True)
+
+    @property
+    def layout(self):
+        """returns the declared layout"""
+        tmp = [r.strip() for r in self._config_parser.layout.splitlines()]
+        if len(tmp) > 1:
+            raise ConanException("Only one layout can be declared in the [layout] section of "
+                                 "the conanfile.txt")
+        return tmp[0] if tmp else None
+
+    @property
+    def requirements(self):
+        """returns a list of requires
+        EX:  "OpenCV/2.4.10@phil/stable"
+        """
+        return [r.strip() for r in self._config_parser.requires.splitlines()]
+
+    @property
+    def tool_requirements(self):
+        """returns a list of tool_requires
+        EX:  "OpenCV/2.4.10@phil/stable"
+        """
+
+        return [r.strip() for r in self._config_parser.tool_requires.splitlines()]
+
+    @property
+    def test_requirements(self):
+        """returns a list of test_requires
+        EX:  "gtest/2.4.10@phil/stable"
+        """
+
+        return [r.strip() for r in self._config_parser.test_requires.splitlines()]
+
+    @property
+    def options(self):
+        return self._config_parser.options
+
+    @property
+    def generators(self):
+        return self._config_parser.generators.splitlines()
