+++ conan/internal/rest/conan_requester.py
@@ -13,7 +13,7 @@ from conan.api.output import ConanOutput
 from conan.internal.cache.home_paths import HomePaths
 
 from conan import __version__
-from conans.client.loader import load_python_file
+from conan.internal.loader import load_python_file
 from conan.internal.errors import scoped_traceback
 from conan.errors import ConanException
