+++ conan/test/utils/test_files.py
@@ -9,7 +9,7 @@ import time
 from io import BytesIO
 
 from conan.tools.files.files import untargz
-from conans.client.subsystems import get_cased_path
+from conan.internal.subsystems import get_cased_path
 from conan.errors import ConanException
 from conan.internal.paths import PACKAGE_TGZ_NAME
 from conans.util.files import gzopen_without_timestamps
