+++ conan/tools/gnu/autotools.py
@@ -2,7 +2,7 @@ import os
 import re
 
 from conan.tools.build import build_jobs, cmd_args_to_string, load_toolchain_args
-from conans.client.subsystems import subsystem_path, deduce_subsystem
+from conan.internal.subsystems import subsystem_path, deduce_subsystem
 from conan.tools.files import chdir
 from conan.tools.microsoft import unix_path
