+++ conans/migrations.py
@@ -2,7 +2,7 @@ import os
 
 from conan import conan_version
 from conan.api.output import ConanOutput
-from conans.client.loader import load_python_file
+from conan.internal.loader import load_python_file
 from conan.errors import ConanException, ConanMigrationError
 from conan.internal.model.version import Version
 from conans.util.files import load, save
