+++ test/unittests/client/conanfile_loader_test.py
@@ -6,7 +6,7 @@ import unittest
 import pytest
 from parameterized import parameterized
 
-from conans.client.loader import ConanFileLoader, ConanFileTextLoader, load_python_file
+from conan.internal.loader import ConanFileLoader, ConanFileTextLoader, load_python_file
 from conan.errors import ConanException
 from conan.test.utils.test_files import temp_folder
 from conans.util.files import save, chdir
