+++ test/unittests/tools/files_patch_test.py
@@ -4,7 +4,7 @@ from textwrap import dedent
 
 from parameterized.parameterized import parameterized
 
-from conans.client.loader import ConanFileLoader
+from conan.internal.loader import ConanFileLoader
 from conan.test.utils.test_files import temp_folder
 from conan.test.utils.tools import TestClient
 from conans.util.files import save, load
